+++ src/api/auth/signUpApi.ts
@@ -1,4 +1,4 @@
-import { SignUpFormData } from "@/src/types/auth/signupFormData";
+import { SignUpFormData } from "@/src/types/auth/signupFormDataType";
 import instance from "../axiosModul";
 
 const withFormData = (data: SignUpFormData) => {
